Use double instead of BigDecimal for VOs
Since we're not going to do operations on them.

+++ backend/src/test/java/ca/bc/gov/backendstartapi/service/parser/ConeAndPollenCountCsvTableParserTest.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 import ca.bc.gov.backendstartapi.exception.CsvTableParsingException;
 import ca.bc.gov.backendstartapi.vo.parser.ConeAndPollenCount;
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 import org.junit.jupiter.api.Test;
@@ -37,8 +36,8 @@ class ConeAndPollenCountCsvTableParserTest {
                     classLoader.getResource("csv/contribution/finalEmptyLine.csv"))));
     assertEquals(
         List.of(
-            new ConeAndPollenCount(1, BigDecimal.valueOf(2.3), BigDecimal.valueOf(4d), 5, 6),
-            new ConeAndPollenCount(7, BigDecimal.valueOf(8.9), BigDecimal.valueOf(10.11), 12, 13)),
+            new ConeAndPollenCount(1, 2.3, 4d, 5, 6),
+            new ConeAndPollenCount(7, 8.9, 10.11, 12, 13)),
         mixes);
   }
