+++ backend/src/test/java/ca/bc/gov/backendstartapi/service/parser/CsvTableRowParserTest.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 import ca.bc.gov.backendstartapi.enums.parser.SmpMixHeader;
 import ca.bc.gov.backendstartapi.exception.CsvTableParsingException;
 import ca.bc.gov.backendstartapi.vo.parser.SmpMixVolume;
-import java.math.BigDecimal;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
@@ -20,14 +19,14 @@ class CsvTableRowParserTest {
     var row =
         parser.parse(
             "1,3.0", List.of(SmpMixHeader.PARENT_TREE_NUMBER, SmpMixHeader.POLLEN_VOLUME_ML));
-    assertEquals(new SmpMixVolume(1, BigDecimal.valueOf(3.0)), row);
+    assertEquals(new SmpMixVolume(1, 3.0), row);
   }
 
   @Test
   void testEmptyCellsValuedAsZero() {
     var row =
         parser.parse("1,", List.of(SmpMixHeader.PARENT_TREE_NUMBER, SmpMixHeader.POLLEN_VOLUME_ML));
-    assertEquals(new SmpMixVolume(1, BigDecimal.valueOf(0d)), row);
+    assertEquals(new SmpMixVolume(1, 0d), row);
   }
 
   @Test
